refactor: move to a separate module
